webapi/tct-batterystatus-w3c-tests/batterystatus/w3c/battery-unplugging-chargingTime-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Test: battery-unplugging-chargingTime</title>
<meta name="flags" content="interact">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #note {
    background-color: #fef1b5;
    border: solid 1px #cdab2d;
    padding: 5px;
    margin: 15px;
    display: block;
  }
</style>

<h2>Description</h2>
<p>
  This test validates that onchargingtimechange and chargingTime attributes exist and
  are set to correct values, with corresponding events fired,
  when the charger is unplugged.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    The device is plugged in to the charger before this test is run.
  </li>
</ol>

<div id="note">
  Unplug the charger and wait for the test to complete.
</div>

<div id="log"></div>

<script>

(function() {

  setup({ explicit_timeout: true });

  var onchargingtimechange_test = async_test(
      'When the device is unplugged in and its charging time is updated, ' +
      'must set the chargingTime attribute\'s value to Infinity and ' +
      'fire a chargingtimechange event.');

  var batterySuccess = function (battery) {
    battery.onchargingtimechange = onchargingtimechange_test.step_func(function () {
      assert_equals(battery.chargingTime, Infinity,
          'The value of the chargingTime attribute must be set to Infinity.');
      onchargingtimechange_test.done();
    });
  };

  var batteryFailure = function (error) {
    onchargingtimechange_test.step(function () {
      assert_unreached(error.message);
    });
    onchargingtimechange_test.done();
  };

  navigator.getBattery().then(batterySuccess, batteryFailure);

})();

</script>
